chore: remove legacy aggregate pipeline
