+++ src/query/service/src/physical_plans/physical_plan_visitor.rs
@@ -157,7 +157,6 @@
 //         Ok(PhysicalPlan::AggregatePartial(AggregatePartial {
 //             plan_id: plan.plan_id,
 //             input: Box::new(input),
-//             enable_experimental_aggregate_hashtable: plan.enable_experimental_aggregate_hashtable,
 //             group_by: plan.group_by.clone(),
 //             group_by_display: plan.group_by_display.clone(),
 //             agg_funcs: plan.agg_funcs.clone(),
