+++ src/query/service/src/pipelines/processors/transforms/aggregator/aggregate_exchange_injector.rs
@@ -14,25 +14,18 @@
 
 use std::sync::Arc;
 
-use databend_common_exception::ErrorCode;
 use databend_common_exception::Result;
-use databend_common_expression::BlockMetaInfoDowncast;
-use databend_common_expression::DataBlock;
 use databend_common_pipeline::core::Pipeline;
 use databend_common_pipeline::core::ProcessorPtr;
 use databend_common_settings::FlightCompression;
-use databend_common_storage::DataOperator;
 
 use crate::physical_plans::AggregateShuffleMode;
 use crate::pipelines::processors::transforms::aggregator::AggregateBucketScatter;
-use crate::pipelines::processors::transforms::aggregator::AggregateMeta;
 use crate::pipelines::processors::transforms::aggregator::AggregateRowScatter;
 use crate::pipelines::processors::transforms::aggregator::AggregatorParams;
 use crate::pipelines::processors::transforms::aggregator::TransformAggregateDeserializer;
 use crate::pipelines::processors::transforms::aggregator::TransformAggregateSerializer;
-use crate::pipelines::processors::transforms::aggregator::TransformAggregateSpillWriter;
 use crate::pipelines::processors::transforms::aggregator::serde::TransformExchangeAggregateSerializer;
-use crate::pipelines::processors::transforms::aggregator::serde::TransformExchangeAsyncBarrier;
 use crate::servers::flight::v1::exchange::DataExchange;
 use crate::servers::flight::v1::exchange::ExchangeInjector;
 use crate::servers::flight::v1::exchange::ExchangeSorting;
@@ -41,52 +34,19 @@ use crate::servers::flight::v1::exchange::ShuffleExchangeParams;
 use crate::servers::flight::v1::scatter::FlightScatter;
 use crate::sessions::QueryContext;
 
-struct AggregateExchangeSorting {}
-
-impl ExchangeSorting for AggregateExchangeSorting {
-    fn block_number(&self, data_block: &DataBlock) -> Result<isize> {
-        match data_block.get_meta() {
-            None => Ok(-1),
-            Some(block_meta_info) => match AggregateMeta::downcast_ref_from(block_meta_info) {
-                None => Err(ErrorCode::Internal(format!(
-                    "Internal error, AggregateExchangeSorting only recv AggregateMeta {:?}",
-                    serde_json::to_string(block_meta_info)
-                ))),
-                Some(meta_info) => Ok(match meta_info {
-                    AggregateMeta::Serialized(payload) => {
-                        payload.max_partition_count as isize * 1000 + payload.bucket
-                    }
-                    AggregateMeta::AggregatePayload(payload) => payload.exchange_block_number(),
-                    AggregateMeta::AggregateSpilling(_)
-                    | AggregateMeta::Spilled(_)
-                    | AggregateMeta::BucketSpilled(_)
-                    | AggregateMeta::NewBucketSpilled(_)
-                    | AggregateMeta::NewSpilled(_) => -1,
-                    AggregateMeta::Partitioned { .. } => {
-                        unreachable!("Partitioned AggregateMeta should not be exchange sorted")
-                    }
-                }),
-            },
-        }
-    }
-}
-
-pub struct AggregateInjector<const ENABLE_EXPERIMENT: bool> {
+pub struct AggregateInjector {
     ctx: Arc<QueryContext>,
     aggregator_params: Arc<AggregatorParams>,
     shuffle_mode: AggregateShuffleMode,
 }
 
-impl<const ENABLE_EXPERIMENT: bool> AggregateInjector<ENABLE_EXPERIMENT> {
+impl AggregateInjector {
     pub fn create(
         ctx: Arc<QueryContext>,
         params: Arc<AggregatorParams>,
         shuffle_mode: AggregateShuffleMode,
-    ) -> Arc<dyn ExchangeInjector>
-    where
-        Self: ExchangeInjector,
-    {
-        Arc::new(AggregateInjector::<ENABLE_EXPERIMENT> {
+    ) -> Arc<dyn ExchangeInjector> {
+        Arc::new(AggregateInjector {
             ctx,
             aggregator_params: params,
             shuffle_mode,
@@ -94,115 +54,7 @@ impl<const ENABLE_EXPERIMENT: bool> AggregateInjector<ENABLE_EXPERIMENT> {
     }
 }
 
-impl ExchangeInjector for AggregateInjector<false> {
-    fn flight_scatter(
-        &self,
-        _: &Arc<QueryContext>,
-        exchange: &DataExchange,
-    ) -> Result<Arc<Box<dyn FlightScatter>>> {
-        match exchange {
-            DataExchange::Merge(_) => unreachable!(),
-            DataExchange::Broadcast(_) => unreachable!(),
-            DataExchange::GlobalShuffleExchange(_) => unreachable!(),
-            DataExchange::NodeToNodeExchange(exchange) => {
-                Ok(Arc::new(Box::new(AggregateRowScatter {
-                    buckets: exchange.destination_ids.len(),
-                    aggregate_params: self.aggregator_params.clone(),
-                })))
-            }
-        }
-    }
-
-    fn exchange_sorting(&self) -> Option<Arc<dyn ExchangeSorting>> {
-        Some(Arc::new(AggregateExchangeSorting {}))
-    }
-
-    fn apply_merge_serializer(
-        &self,
-        _: &MergeExchangeParams,
-        _compression: Option<FlightCompression>,
-        pipeline: &mut Pipeline,
-    ) -> Result<()> {
-        let params = self.aggregator_params.clone();
-
-        let operator = DataOperator::instance().spill_operator();
-        let location_prefix = self.ctx.query_id_spill_prefix();
-
-        pipeline.add_transform(|input, output| {
-            Ok(ProcessorPtr::create(
-                TransformAggregateSpillWriter::try_create(
-                    self.ctx.clone(),
-                    input,
-                    output,
-                    operator.clone(),
-                    params.clone(),
-                    location_prefix.clone(),
-                )?,
-            ))
-        })?;
-
-        pipeline.add_transform(|input, output| {
-            TransformAggregateSerializer::try_create(input, output, params.clone())
-        })
-    }
-
-    fn apply_shuffle_serializer(
-        &self,
-        shuffle_params: &ShuffleExchangeParams,
-        compression: Option<FlightCompression>,
-        pipeline: &mut Pipeline,
-    ) -> Result<()> {
-        let params = self.aggregator_params.clone();
-        let operator = DataOperator::instance().spill_operator();
-        let location_prefix = self.ctx.query_id_spill_prefix();
-
-        let local_id = &shuffle_params.executor_id;
-        let local_pos = shuffle_params
-            .destination_ids
-            .iter()
-            .position(|x| x == local_id)
-            .unwrap();
-
-        pipeline.add_transform(|input, output| {
-            Ok(ProcessorPtr::create(
-                TransformExchangeAggregateSerializer::try_create(
-                    self.ctx.clone(),
-                    input,
-                    output,
-                    operator.clone(),
-                    location_prefix.clone(),
-                    params.clone(),
-                    compression,
-                    local_pos,
-                )?,
-            ))
-        })?;
-        pipeline.add_transform(TransformExchangeAsyncBarrier::try_create)
-    }
-
-    fn apply_merge_deserializer(
-        &self,
-        params: &MergeExchangeParams,
-        pipeline: &mut Pipeline,
-    ) -> Result<()> {
-        pipeline.add_transform(|input, output| {
-            TransformAggregateDeserializer::try_create(input, output, &params.schema)
-        })
-    }
-
-    fn apply_shuffle_deserializer(
-        &self,
-        params: &ShuffleExchangeParams,
-        pipeline: &mut Pipeline,
-    ) -> Result<()> {
-        pipeline.add_transform(|input, output| {
-            TransformAggregateDeserializer::try_create(input, output, &params.schema)
-        })?;
-        Ok(())
-    }
-}
-
-impl ExchangeInjector for AggregateInjector<true> {
+impl ExchangeInjector for AggregateInjector {
     fn flight_scatter(
         &self,
         _: &Arc<QueryContext>,
@@ -248,8 +100,6 @@ impl ExchangeInjector for AggregateInjector<true> {
         pipeline: &mut Pipeline,
     ) -> Result<()> {
         let params = self.aggregator_params.clone();
-        let operator = DataOperator::instance().spill_operator();
-        let location_prefix = self.ctx.query_id_spill_prefix();
 
         let local_id = &shuffle_params.executor_id;
         let local_pos = shuffle_params
@@ -264,8 +114,6 @@ impl ExchangeInjector for AggregateInjector<true> {
                     self.ctx.clone(),
                     input,
                     output,
-                    operator.clone(),
-                    location_prefix.clone(),
                     params.clone(),
                     compression,
                     local_pos,
@@ -273,8 +121,7 @@ impl ExchangeInjector for AggregateInjector<true> {
             ))
         })?;
 
-        // TODO: this could be removed when experiment aggregate stabilize
-        pipeline.add_transform(TransformExchangeAsyncBarrier::try_create)
+        Ok(())
     }
 
     fn apply_merge_deserializer(
