+++ src/query/service/src/pipelines/processors/transforms/aggregator/build_partition_bucket.rs
@@ -23,27 +23,22 @@ use databend_common_pipeline::core::PipeItem;
 use databend_common_pipeline::core::Pipeline;
 use databend_common_pipeline::core::ProcessorPtr;
 use databend_common_pipeline::core::TransformPipeBuilder;
-use databend_common_storage::DataOperator;
-use tokio::sync::Semaphore;
 
 use crate::clusters::ClusterHelper;
 use crate::physical_plans::AggregateShuffleMode;
 use crate::pipelines::processors::transforms::aggregator::AggregateBucketScatter;
 use crate::pipelines::processors::transforms::aggregator::AggregateRowScatter;
+use crate::pipelines::processors::transforms::aggregator::AggregateSpillReader;
 use crate::pipelines::processors::transforms::aggregator::AggregatorParams;
 use crate::pipelines::processors::transforms::aggregator::LocalScatterTransform;
-use crate::pipelines::processors::transforms::aggregator::NewAggregateSpillReader;
-use crate::pipelines::processors::transforms::aggregator::NewTransformFinalAggregate;
 use crate::pipelines::processors::transforms::aggregator::RowShuffleReaderTransform;
-use crate::pipelines::processors::transforms::aggregator::TransformAggregateSpillReader;
 use crate::pipelines::processors::transforms::aggregator::TransformFinalAggregate;
-use crate::pipelines::processors::transforms::aggregator::transform_partition_bucket::TransformPartitionBucket;
 use crate::servers::flight::v1::exchange::ExchangeShuffleTransform;
 use crate::sessions::QueryContext;
 use crate::sessions::TableContextCluster;
 use crate::sessions::TableContextSettings;
 
-fn build_partition_bucket_experimental(
+pub fn build_partition_bucket(
     pipeline: &mut Pipeline,
     params: Arc<AggregatorParams>,
     after_worker: usize,
@@ -59,7 +54,7 @@ fn build_partition_bucket_experimental(
                 Ok(ProcessorPtr::create(RowShuffleReaderTransform::create(
                     input,
                     output,
-                    NewAggregateSpillReader::try_create(ctx.clone(), schema.clone())?,
+                    AggregateSpillReader::try_create(ctx.clone(), schema.clone())?,
                 )))
             })?;
 
@@ -114,7 +109,7 @@ fn build_partition_bucket_experimental(
     for id in 0..final_parallelism {
         let input_port = InputPort::create();
         let output_port = OutputPort::create();
-        let processor = NewTransformFinalAggregate::try_create(
+        let processor = TransformFinalAggregate::try_create(
             input_port.clone(),
             output_port.clone(),
             params.clone(),
@@ -133,59 +128,4 @@ fn build_partition_bucket_experimental(
     pipeline.resize(after_worker, true)?;
 
     Ok(())
-}
-
-fn build_partition_bucket_legacy(
-    pipeline: &mut Pipeline,
-    params: Arc<AggregatorParams>,
-    max_restore_worker: u64,
-    after_worker: usize,
-) -> Result<()> {
-    let operator = DataOperator::instance().spill_operator();
-
-    let input_num = pipeline.output_len();
-    let transform = TransformPartitionBucket::create(input_num, params.clone())?;
-
-    let output = transform.get_output();
-    let inputs_port = transform.get_inputs();
-
-    pipeline.add_pipe(Pipe::create(inputs_port.len(), 1, vec![PipeItem::create(
-        ProcessorPtr::create(Box::new(transform)),
-        inputs_port,
-        vec![output],
-    )]));
-
-    pipeline.try_resize(std::cmp::min(input_num, max_restore_worker as usize))?;
-    let semaphore = Arc::new(Semaphore::new(params.max_spill_io_requests));
-    pipeline.add_transform(|input, output| {
-        let operator = operator.clone();
-        TransformAggregateSpillReader::create(input, output, operator, semaphore.clone())
-    })?;
-    pipeline.add_transform(|input, output| {
-        Ok(ProcessorPtr::create(TransformFinalAggregate::try_create(
-            input,
-            output,
-            params.clone(),
-        )?))
-    })?;
-    pipeline.try_resize(after_worker)?;
-
-    Ok(())
-}
-
-/// Build partition bucket pipeline based on the experiment_aggregate flag.
-/// Dispatches to either experimental or legacy implementation.
-pub fn build_partition_bucket(
-    pipeline: &mut Pipeline,
-    params: Arc<AggregatorParams>,
-    max_restore_worker: u64,
-    after_worker: usize,
-    ctx: Arc<QueryContext>,
-    shuffle_mode: AggregateShuffleMode,
-) -> Result<()> {
-    if params.enable_experiment_aggregate {
-        build_partition_bucket_experimental(pipeline, params, after_worker, ctx, shuffle_mode)
-    } else {
-        build_partition_bucket_legacy(pipeline, params, max_restore_worker, after_worker)
-    }
 }
