+++ src/query/service/src/pipelines/processors/transforms/aggregator/mod.rs
@@ -14,26 +14,26 @@
 
 mod aggregate_exchange_injector;
 mod aggregate_meta;
+mod aggregate_spiller;
 mod aggregator_params;
 mod build_partition_bucket;
-mod new_aggregate;
+mod row_shuffle_reader_transform;
 mod serde;
 mod statistics;
 mod transform_aggregate_expand;
 mod transform_aggregate_final;
 mod transform_aggregate_partial;
-mod transform_partition_bucket;
 mod transform_single_key;
 
 pub use aggregate_exchange_injector::AggregateInjector;
 pub use aggregate_meta::*;
+pub use aggregate_spiller::*;
 pub use aggregator_params::AggregatorParams;
 pub use build_partition_bucket::build_partition_bucket;
 pub use databend_common_expression::aggregate::AggregatePayload;
-pub use databend_common_expression::aggregate::BucketSpilledPayload;
 pub use databend_common_expression::aggregate::SerializedPayload;
 pub use databend_query_script_udf_support::create_udaf_script_function;
-pub use new_aggregate::*;
+pub use row_shuffle_reader_transform::RowShuffleReaderTransform;
 pub use transform_aggregate_expand::TransformExpandGroupingSets;
 pub use transform_aggregate_final::TransformFinalAggregate;
 pub use transform_aggregate_partial::TransformPartialAggregate;
