--- src/query/service/src/pipelines/processors/transforms/aggregator/new_aggregate/mod.rs
@@ -1,27 +1,0 @@
-// Copyright 2021 Datafuse Labs
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-mod new_aggregate_spiller;
-mod new_transform_aggregate_partial;
-mod new_transform_final_aggregate;
-mod row_shuffle_reader_transform;
-
-pub use new_aggregate_spiller::LocalPartitionStream;
-pub use new_aggregate_spiller::NewAggregateSpillReader;
-pub use new_aggregate_spiller::NewAggregateSpiller;
-pub use new_aggregate_spiller::PartitionStream;
-pub use new_aggregate_spiller::SharedPartitionStream;
-pub use new_transform_aggregate_partial::NewTransformPartialAggregate;
-pub use new_transform_final_aggregate::NewTransformFinalAggregate;
-pub use row_shuffle_reader_transform::RowShuffleReaderTransform;
