+++ src/query/service/src/pipelines/processors/transforms/aggregator/serde/transform_aggregate_serializer.rs
@@ -13,22 +13,17 @@
 // limitations under the License.
 
 use std::any::Any;
-use std::fmt::Formatter;
 use std::sync::Arc;
 
 use databend_common_exception::Result;
-use databend_common_expression::BlockMetaInfo;
 use databend_common_expression::BlockMetaInfoDowncast;
-use databend_common_expression::BlockMetaInfoPtr;
 use databend_common_expression::DataBlock;
 use databend_common_expression::PayloadFlushState;
-use databend_common_expression::local_block_meta_serde;
 use databend_common_pipeline::core::Event;
 use databend_common_pipeline::core::InputPort;
 use databend_common_pipeline::core::OutputPort;
 use databend_common_pipeline::core::Processor;
 use databend_common_pipeline::core::ProcessorPtr;
-use futures::future::BoxFuture;
 
 use crate::pipelines::processors::transforms::aggregator::AggregateMeta;
 use crate::pipelines::processors::transforms::aggregator::AggregatePayload;
@@ -132,36 +127,6 @@ impl TransformAggregateSerializer {
         Ok(Event::Sync)
     }
 }
-
-pub enum FlightSerialized {
-    DataBlock(DataBlock),
-    Future(BoxFuture<'static, Result<DataBlock>>),
-}
-
-unsafe impl Sync for FlightSerialized {}
-
-pub struct FlightSerializedMeta {
-    pub serialized_blocks: Vec<FlightSerialized>,
-}
-
-impl FlightSerializedMeta {
-    pub fn create(blocks: Vec<FlightSerialized>) -> BlockMetaInfoPtr {
-        Box::new(FlightSerializedMeta {
-            serialized_blocks: blocks,
-        })
-    }
-}
-
-impl std::fmt::Debug for FlightSerializedMeta {
-    fn fmt(&self, f: &mut Formatter) -> std::fmt::Result {
-        f.debug_struct("FlightSerializedMeta").finish()
-    }
-}
-
-local_block_meta_serde!(FlightSerializedMeta);
-
-#[typetag::serde(name = "exchange_shuffle")]
-impl BlockMetaInfo for FlightSerializedMeta {}
 
 pub struct SerializeAggregateStream {
     _params: Arc<AggregatorParams>,
