+++ src/query/service/src/pipelines/processors/transforms/aggregator/serde/transform_deserializer.rs
@@ -21,11 +21,9 @@ use databend_common_expression::BlockMetaInfoDowncast;
 use databend_common_expression::DataBlock;
 use databend_common_expression::DataSchemaRef;
 use databend_common_expression::types::AccessType;
-use databend_common_expression::types::ArrayType;
 use databend_common_expression::types::BinaryType;
 use databend_common_expression::types::NumberType;
 use databend_common_expression::types::StringType;
-use databend_common_expression::types::UInt64Type;
 use databend_common_io::prelude::BinaryRead;
 use databend_common_io::prelude::bincode_deserialize_from_slice;
 use databend_common_pipeline::core::InputPort;
@@ -38,13 +36,11 @@ use databend_common_storages_parquet::deserialize_row_group_meta_from_bytes;
 use crate::pipelines::processors::transforms::aggregator::AggregateMeta;
 use crate::pipelines::processors::transforms::aggregator::AggregateSerdeMeta;
 use crate::pipelines::processors::transforms::aggregator::BUCKET_TYPE;
-use crate::pipelines::processors::transforms::aggregator::BucketSpilledPayload;
-use crate::pipelines::processors::transforms::aggregator::NEW_SPILLED_TYPE;
-use crate::pipelines::processors::transforms::aggregator::NewSpilledPayload;
 use crate::pipelines::processors::transforms::aggregator::PARTITIONED_AGGREGATE_TYPE;
 use crate::pipelines::processors::transforms::aggregator::PartitionedData;
 use crate::pipelines::processors::transforms::aggregator::SPILLED_TYPE;
 use crate::pipelines::processors::transforms::aggregator::SerializedPayload;
+use crate::pipelines::processors::transforms::aggregator::SpilledPayload;
 use crate::pipelines::processors::transforms::aggregator::exchange_defines;
 use crate::servers::flight::v1::exchange::serde::ExchangeDeserializeMeta;
 use crate::servers::flight::v1::exchange::serde::deserialize_block;
@@ -189,48 +185,6 @@ impl TransformDeserializer {
                     .collect::<Vec<_>>();
 
                 let buckets = NumberType::<i64>::try_downcast_column(&columns[0]).unwrap();
-                let data_range_start = NumberType::<u64>::try_downcast_column(&columns[1]).unwrap();
-                let data_range_end = NumberType::<u64>::try_downcast_column(&columns[2]).unwrap();
-                let columns_layout =
-                    ArrayType::<UInt64Type>::try_downcast_column(&columns[3]).unwrap();
-
-                let columns_layout_data = columns_layout.values().as_slice();
-                let columns_layout_offsets = columns_layout.offsets();
-
-                let mut buckets_payload = Vec::with_capacity(data_block.num_rows());
-                for index in 0..data_block.num_rows() {
-                    unsafe {
-                        buckets_payload.push(BucketSpilledPayload {
-                            bucket: *buckets.get_unchecked(index) as isize,
-                            location: meta.location.clone().unwrap(),
-                            data_range: *data_range_start.get_unchecked(index)
-                                ..*data_range_end.get_unchecked(index),
-                            columns_layout: columns_layout_data[columns_layout_offsets[index]
-                                as usize
-                                ..columns_layout_offsets[index + 1] as usize]
-                                .to_vec(),
-                            max_partition_count: meta.max_partition_count,
-                        });
-                    }
-                }
-
-                Ok(vec![DataBlock::empty_with_meta(
-                    AggregateMeta::create_spilled(buckets_payload),
-                )])
-            }
-            NEW_SPILLED_TYPE => {
-                let data_schema = Arc::new(exchange_defines::new_spilled_schema());
-                let arrow_schema = Arc::new(exchange_defines::new_spilled_arrow_schema());
-                let data_block =
-                    deserialize_block(dict, fragment_data, &data_schema, arrow_schema.clone())?;
-
-                let columns = data_block
-                    .columns()
-                    .iter()
-                    .map(|c| c.as_column().unwrap().clone())
-                    .collect::<Vec<_>>();
-
-                let buckets = NumberType::<i64>::try_downcast_column(&columns[0]).unwrap();
                 let locations = StringType::try_downcast_column(&columns[1]).unwrap();
                 let row_groups = BinaryType::try_downcast_column(&columns[2]).unwrap();
 
@@ -242,7 +196,7 @@ impl TransformDeserializer {
                         let row_group_bytes = row_groups.index_unchecked(index);
                         let row_group = deserialize_row_group_meta_from_bytes(row_group_bytes)?;
 
-                        spilled_payloads.push(NewSpilledPayload {
+                        spilled_payloads.push(SpilledPayload {
                             bucket,
                             location,
                             row_group,
@@ -256,7 +210,7 @@ impl TransformDeserializer {
                 } else {
                     let partitioned = AggregateMeta::create_partitioned(
                         None,
-                        PartitionedData::NewBucketSpilled(spilled_payloads),
+                        PartitionedData::BucketSpilled(spilled_payloads),
                     );
                     return Ok(vec![DataBlock::empty_with_meta(partitioned)]);
                 }
