+++ src/query/service/src/pipelines/processors/transforms/aggregator/transform_aggregate_final.rs
@@ -12,84 +12,331 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use std::any::Any;
+use std::mem;
 use std::sync::Arc;
+use std::sync::atomic::AtomicU64;
+use std::sync::atomic::Ordering;
 
+use async_channel::Receiver;
+use async_channel::Sender;
 use bumpalo::Bump;
+use databend_common_exception::ErrorCode;
 use databend_common_exception::Result;
 use databend_common_expression::AggregateHashTable;
-use databend_common_expression::AggregatePayload;
+use databend_common_expression::BlockMetaInfoDowncast;
 use databend_common_expression::DataBlock;
 use databend_common_expression::HashTableConfig;
 use databend_common_expression::PayloadFlushState;
-use databend_common_expression::SerializedPayload;
+use databend_common_pipeline::core::Event;
 use databend_common_pipeline::core::InputPort;
 use databend_common_pipeline::core::OutputPort;
 use databend_common_pipeline::core::Processor;
-use databend_common_pipeline_transforms::processors::BlockMetaTransform;
-use databend_common_pipeline_transforms::processors::BlockMetaTransformer;
+use databend_common_pipeline_transforms::MemorySettings;
 
+use crate::pipelines::memory_settings::MemorySettingsExt;
 use crate::pipelines::processors::transforms::aggregator::AggregateMeta;
+use crate::pipelines::processors::transforms::aggregator::AggregatePayload;
+use crate::pipelines::processors::transforms::aggregator::AggregateSpiller;
 use crate::pipelines::processors::transforms::aggregator::AggregatorParams;
-use crate::pipelines::processors::transforms::aggregator::PartitionItem;
+use crate::pipelines::processors::transforms::aggregator::LocalPartitionStream;
 use crate::pipelines::processors::transforms::aggregator::PartitionedData;
+use crate::pipelines::processors::transforms::aggregator::SerializedPayload;
+use crate::pipelines::processors::transforms::aggregator::SpilledPayload;
+use crate::pipelines::processors::transforms::aggregator::statistics::AggregationStatistics;
+use crate::pipelines::processors::transforms::aggregator::transform_aggregate_partial::HashTable;
+use crate::sessions::QueryContext;
+use crate::sessions::TableContextSettings;
+
+const SPILL_BUCKET_NUM: usize = 4;
+const SPILL_BUCKET_BITS: u64 = SPILL_BUCKET_NUM.trailing_zeros() as u64;
+
+enum Stage {
+    Input,
+    Channel,
+}
+
+pub struct FinalAggregateTask {
+    task_id: u64,
+    spilled_depth: usize,
+    spilled_payload: Vec<SpilledPayload>,
+    tx: Sender<FinalAggregateTask>,
+}
 
 pub struct TransformFinalAggregate {
+    input: Arc<InputPort>,
+    output: Arc<OutputPort>,
+    input_data: Option<AggregateMeta>,
+    channel_data: Option<FinalAggregateTask>,
+    should_finish: bool,
+    tx: Option<Sender<FinalAggregateTask>>,
+    rx: Receiver<FinalAggregateTask>,
+    stage: Stage,
+    spilled_occurred: bool,
+
+    hashtable: HashTable,
     params: Arc<AggregatorParams>,
     flush_state: PayloadFlushState,
+    statistics: AggregationStatistics,
+    _id: usize,
+    base_consumed_bits: u64,
+    max_partition_depth: usize,
+    current_partition_depth: usize,
+    spiller: AggregateSpiller<LocalPartitionStream>,
+    settings: MemorySettings,
+    max_aggregate_spill_level: usize,
+    next_task_id: Arc<AtomicU64>,
 }
 
 impl TransformFinalAggregate {
     pub fn try_create(
         input: Arc<InputPort>,
         output: Arc<OutputPort>,
-
         params: Arc<AggregatorParams>,
+        _id: usize,
+        base_consumed_bits: u64,
+        ctx: Arc<QueryContext>,
+        tx: Sender<FinalAggregateTask>,
+        rx: Receiver<FinalAggregateTask>,
+        next_task_id: Arc<AtomicU64>,
     ) -> Result<Box<dyn Processor>> {
-        Ok(BlockMetaTransformer::create(
+        let settings = ctx.get_settings();
+        let available_partition_depths =
+            (48_u64.saturating_sub(base_consumed_bits) / SPILL_BUCKET_BITS) as usize;
+        let max_partition_depth = available_partition_depths.saturating_sub(1);
+        let max_aggregate_spill_level =
+            (settings.get_max_aggregate_spill_level()? as usize).min(max_partition_depth);
+
+        let hashtable = Self::create_hashtable(&params, base_consumed_bits, 0);
+        let flush_state = PayloadFlushState::default();
+
+        let spiller = AggregateSpiller::try_create(
+            ctx.clone(),
+            SPILL_BUCKET_NUM,
+            params.spill_schema(),
+            LocalPartitionStream::new(0, params.max_block_bytes, SPILL_BUCKET_NUM),
+        )?;
+
+        Ok(Box::new(TransformFinalAggregate {
             input,
             output,
-            TransformFinalAggregate {
-                params,
-                flush_state: PayloadFlushState::default(),
-            },
-        ))
+            input_data: None,
+            channel_data: None,
+            should_finish: false,
+            tx: Some(tx),
+            rx,
+            stage: Stage::Input,
+            spilled_occurred: false,
+            hashtable: HashTable::AggregateHashTable(hashtable),
+            params,
+            flush_state,
+            statistics: AggregationStatistics::new("FinalAggregate"),
+            _id,
+            base_consumed_bits,
+            max_partition_depth,
+            current_partition_depth: 0,
+            spiller,
+            settings: MemorySettings::from_aggregate_settings(&ctx)?,
+            max_aggregate_spill_level,
+            next_task_id,
+        }))
+    }
+}
+
+impl TransformFinalAggregate {
+    fn next_task_id(&self) -> u64 {
+        self.next_task_id.fetch_add(1, Ordering::Relaxed)
     }
 
-    fn transform_agg_hashtable(
+    fn partition_start_bit(base_consumed_bits: u64, spilled_depth: usize) -> u64 {
+        base_consumed_bits + spilled_depth as u64 * SPILL_BUCKET_BITS
+    }
+
+    fn create_hashtable(
+        params: &Arc<AggregatorParams>,
+        base_consumed_bits: u64,
+        spilled_depth: usize,
+    ) -> AggregateHashTable {
+        AggregateHashTable::new(
+            params.group_data_types.clone(),
+            params.aggregate_functions.clone(),
+            HashTableConfig::default()
+                .with_initial_radix_bits(SPILL_BUCKET_BITS)
+                .with_partition_start_bit(Self::partition_start_bit(
+                    base_consumed_bits,
+                    spilled_depth,
+                )),
+            Arc::new(Bump::new()),
+        )
+    }
+
+    fn reset_hashtable(&mut self, spilled_depth: usize) {
+        let partition_depth = spilled_depth.min(self.max_partition_depth);
+        self.hashtable = HashTable::AggregateHashTable(Self::create_hashtable(
+            &self.params,
+            self.base_consumed_bits,
+            partition_depth,
+        ));
+        self.current_partition_depth = partition_depth;
+    }
+
+    // One final-aggregate processor handles both the original input stream and
+    // recursively spilled tasks received from the channel. Those tasks may
+    // belong to different spill depths, and each depth must consume a different
+    // hash-bit window when building the internal partitions. The window start is
+    // `base_consumed_bits + spilled_depth * SPILL_BUCKET_BITS`.
+    fn ensure_spill_depth(&mut self, spilled_depth: usize) {
+        let partition_depth = spilled_depth.min(self.max_partition_depth);
+        if self.current_partition_depth != partition_depth {
+            self.reset_hashtable(spilled_depth);
+        }
+    }
+
+    fn handle_serialized(
         &mut self,
-        bucket: isize,
-        data: PartitionedData,
-    ) -> Result<DataBlock> {
-        let mut agg_hashtable = None;
-        match data {
-            PartitionedData::Empty => {}
-            PartitionedData::Serialized(payloads) => {
-                for payload in payloads {
-                    self.combine_serialized_payload(&mut agg_hashtable, bucket, payload)?;
-                }
-            }
-            PartitionedData::AggregatePayload(payloads) => {
-                for payload in payloads {
-                    self.combine_aggregate_payload(&mut agg_hashtable, bucket, payload)?;
-                }
-            }
-            PartitionedData::Mixed(items) => {
-                for item in items {
-                    match item {
-                        PartitionItem::Serialized(payload) => {
-                            self.combine_serialized_payload(&mut agg_hashtable, bucket, payload)?;
-                        }
-                        PartitionItem::AggregatePayload(payload) => {
-                            self.combine_aggregate_payload(&mut agg_hashtable, bucket, payload)?;
-                        }
-                        _ => unreachable!("unexpected partitioned aggregate data"),
-                    }
-                }
-            }
-            _ => unreachable!("unexpected partitioned aggregate data"),
+        payload: SerializedPayload,
+        need_check_spill: bool,
+    ) -> Result<()> {
+        if payload.data_block.is_empty() {
+            return Ok(());
         }
 
-        if let Some(mut ht) = agg_hashtable {
+        let rows = payload.data_block.num_rows();
+        let bytes = payload.data_block.memory_size();
+        self.statistics.record_block(rows, bytes);
+
+        let partitioned_payload = payload.convert_to_partitioned_payload(
+            self.params.group_data_types.clone(),
+            self.params.aggregate_functions.clone(),
+            self.params.num_states(),
+            0,
+            Arc::new(Bump::new()),
+        )?;
+
+        if let HashTable::AggregateHashTable(ht) = &mut self.hashtable {
+            ht.combine_payloads(&partitioned_payload, &mut self.flush_state)?;
+        }
+
+        self.check_spill(need_check_spill)
+    }
+
+    fn handle_aggregate_payload(
+        &mut self,
+        payload: AggregatePayload,
+        need_check_spill: bool,
+    ) -> Result<()> {
+        let rows = payload.payload.len();
+        let bytes = payload.payload.memory_size();
+        self.statistics.record_block(rows, bytes);
+
+        if let HashTable::AggregateHashTable(ht) = &mut self.hashtable {
+            ht.combine_payload(&payload.payload, &mut self.flush_state)?;
+        }
+
+        self.check_spill(need_check_spill)
+    }
+
+    fn check_spill(&mut self, need_check_spill: bool) -> Result<()> {
+        // If already trigger spilled for this task, we continue to spill the remaining part
+        if self.spilled_occurred || (need_check_spill && self.settings.check_spill()) {
+            self.spill_out()?;
+        }
+
+        Ok(())
+    }
+
+    fn handle_meta(&mut self, meta: AggregateMeta, need_check_spill: bool) -> Result<()> {
+        match meta {
+            AggregateMeta::Serialized(payload) => {
+                self.handle_serialized(payload, need_check_spill)?;
+            }
+            AggregateMeta::AggregatePayload(payload) => {
+                self.handle_aggregate_payload(payload, need_check_spill)?;
+            }
+            AggregateMeta::Spilled(payloads) => {
+                for payload in payloads {
+                    let restored = self.spiller.restore(payload)?;
+                    self.handle_serialized(restored, need_check_spill)?;
+                }
+            }
+            AggregateMeta::BucketSpilled(payload) => {
+                let restored = self.spiller.restore(payload)?;
+                self.handle_serialized(restored, need_check_spill)?;
+            }
+            AggregateMeta::Partitioned { data, .. } => match data {
+                PartitionedData::Empty => {}
+                PartitionedData::Serialized(payloads) => {
+                    for payload in payloads {
+                        self.handle_serialized(payload, need_check_spill)?;
+                    }
+                }
+                PartitionedData::AggregatePayload(payloads) => {
+                    for payload in payloads {
+                        self.handle_aggregate_payload(payload, need_check_spill)?;
+                    }
+                }
+                PartitionedData::BucketSpilled(payloads) => {
+                    for payload in payloads {
+                        let restored = self.spiller.restore(payload)?;
+                        self.handle_serialized(restored, need_check_spill)?;
+                    }
+                }
+                PartitionedData::Mixed(items) => {
+                    for item in items {
+                        self.handle_meta(AggregateMeta::from(item), need_check_spill)?;
+                    }
+                }
+            },
+        }
+        Ok(())
+    }
+
+    fn spill_out(&mut self) -> Result<()> {
+        self.spilled_occurred = true;
+        if let HashTable::AggregateHashTable(v) = mem::take(&mut self.hashtable) {
+            for (bucket, payload) in v.payload.into_non_empty_bucket_payloads() {
+                let data_block = payload.aggregate_flush_all()?.consume_convert_to_full();
+                self.spiller.spill(bucket, data_block)?;
+            }
+        } else {
+            unreachable!("[TRANSFORM-AGGREGATOR] Invalid hash table state during spill check")
+        }
+        self.reset_hashtable(self.current_partition_depth);
+        Ok(())
+    }
+
+    fn finish(
+        &mut self,
+        task_id: Option<u64>,
+        spilled_depth: usize,
+        tx: Sender<FinalAggregateTask>,
+    ) -> Result<()> {
+        if self.spilled_occurred {
+            let (output_rows, hash_index_resizes) = match &self.hashtable {
+                HashTable::AggregateHashTable(ht) => {
+                    (ht.payload.len(), ht.hash_index_resize_count())
+                }
+                _ => unreachable!("[TRANSFORM-AGGREGATOR] Invalid hash table state before spill"),
+            };
+            self.spill_finish(spilled_depth, tx)?;
+            if let Some(task_id) = task_id {
+                self.statistics.log_task_finish_statistics(
+                    task_id,
+                    self._id,
+                    spilled_depth,
+                    output_rows,
+                    hash_index_resizes,
+                    true,
+                );
+            } else {
+                self.statistics.reset();
+            }
+
+            self.spilled_occurred = false;
+            return Ok(());
+        }
+
+        if let HashTable::AggregateHashTable(mut ht) = mem::take(&mut self.hashtable) {
             let mut blocks = vec![];
             self.flush_state.clear();
 
@@ -105,89 +352,172 @@ impl TransformFinalAggregate {
                 }
             }
 
-            if blocks.is_empty() {
-                return Ok(self.params.empty_result_block());
+            if !blocks.is_empty() {
+                let concat = DataBlock::concat(&blocks)?;
+                if !concat.is_empty() {
+                    self.output.push_data(Ok(concat));
+                }
             }
-            return DataBlock::concat(&blocks);
+            if let Some(task_id) = task_id {
+                self.statistics.log_task_finish_statistics(
+                    task_id,
+                    self._id,
+                    spilled_depth,
+                    ht.payload.len(),
+                    ht.hash_index_resize_count(),
+                    false,
+                );
+            } else {
+                self.statistics.log_finish_statistics(&ht);
+            }
+            self.reset_hashtable(self.current_partition_depth);
         }
 
-        Ok(self.params.empty_result_block())
-    }
-
-    fn combine_serialized_payload(
-        &mut self,
-        agg_hashtable: &mut Option<AggregateHashTable>,
-        bucket: isize,
-        payload: SerializedPayload,
-    ) -> Result<()> {
-        match agg_hashtable.as_mut() {
-            Some(ht) => {
-                debug_assert!(bucket == payload.bucket);
-
-                let payload = payload.convert_to_partitioned_payload(
-                    self.params.group_data_types.clone(),
-                    self.params.aggregate_functions.clone(),
-                    self.params.num_states(),
-                    0,
-                    Arc::new(Bump::new()),
-                )?;
-                ht.combine_payloads(&payload, &mut self.flush_state)?;
-            }
-            None => {
-                debug_assert!(bucket == payload.bucket);
-                *agg_hashtable = Some(payload.convert_to_aggregate_table(
-                    self.params.group_data_types.clone(),
-                    self.params.aggregate_functions.clone(),
-                    self.params.num_states(),
-                    0,
-                    Arc::new(Bump::new()),
-                    true,
-                )?);
-            }
-        }
         Ok(())
     }
 
-    fn combine_aggregate_payload(
-        &mut self,
-        agg_hashtable: &mut Option<AggregateHashTable>,
-        bucket: isize,
-        payload: AggregatePayload,
-    ) -> Result<()> {
-        match agg_hashtable.as_mut() {
-            Some(ht) => {
-                debug_assert!(bucket == payload.bucket);
-                ht.combine_payload(&payload.payload, &mut self.flush_state)?;
-            }
-            None => {
-                debug_assert!(bucket == payload.bucket);
-                let capacity = AggregateHashTable::get_capacity_for_count(payload.payload.len());
-                let mut hashtable = AggregateHashTable::new_with_capacity(
-                    self.params.group_data_types.clone(),
-                    self.params.aggregate_functions.clone(),
-                    HashTableConfig::default().with_initial_radix_bits(0),
-                    capacity,
-                    Arc::new(Bump::new()),
-                );
-                hashtable.combine_payload(&payload.payload, &mut self.flush_state)?;
-                *agg_hashtable = Some(hashtable);
-            }
+    fn spill_finish(&mut self, spilled_depth: usize, tx: Sender<FinalAggregateTask>) -> Result<()> {
+        self.spill_out()?;
+
+        let spilled_payload = self.spiller.spill_finish()?;
+        let mut chunks = (0..SPILL_BUCKET_NUM).map(|_| vec![]).collect::<Vec<_>>();
+        for payload in spilled_payload.into_iter() {
+            chunks[payload.bucket as usize].push(payload);
         }
+
+        let next_spill_depth = spilled_depth + 1;
+        for (bucket, chunk) in chunks.into_iter().enumerate() {
+            let task_id = self.next_task_id();
+            let rows = chunk
+                .iter()
+                .map(|payload| payload.row_group.num_rows() as usize)
+                .sum::<usize>();
+            log::info!(
+                "Spill finish emitted task: task_id={}, processor={}, spill_depth={}, bucket={}, payloads={}, rows={}",
+                task_id,
+                self._id,
+                spilled_depth,
+                bucket,
+                chunk.len(),
+                rows,
+            );
+            let spilled = FinalAggregateTask {
+                task_id,
+                spilled_depth: next_spill_depth,
+                spilled_payload: chunk,
+                tx: tx.clone(),
+            };
+            tx.try_send(spilled).map_err(|e| {
+                ErrorCode::Internal(format!(
+                    "Failed to send final aggregate meta to spiller: {}",
+                    e
+                ))
+            })?;
+        }
+
         Ok(())
     }
 }
 
-impl BlockMetaTransform<AggregateMeta> for TransformFinalAggregate {
-    const NAME: &'static str = "TransformFinalAggregate";
+#[async_trait::async_trait]
+impl Processor for TransformFinalAggregate {
+    fn name(&self) -> String {
+        "TransformFinalAggregate".to_string()
+    }
 
-    fn transform(&mut self, meta: AggregateMeta) -> Result<Vec<DataBlock>> {
-        let AggregateMeta::Partitioned {
-            bucket: Some(bucket),
-            data,
-        } = meta
-        else {
-            unreachable!("TransformFinalAggregate only recv partitioned aggregate meta")
-        };
-        Ok(vec![self.transform_agg_hashtable(bucket, data)?])
+    fn as_any(&mut self) -> &mut dyn Any {
+        self
+    }
+
+    fn event(&mut self) -> Result<Event> {
+        if self.output.is_finished() {
+            let _ = self.tx.take();
+            self.input.finish();
+            return Ok(Event::Finished);
+        }
+
+        if !self.output.can_push() {
+            self.input.set_not_need_data();
+            return Ok(Event::NeedConsume);
+        }
+
+        if self.input_data.is_some() {
+            return Ok(Event::Sync);
+        }
+
+        if self.channel_data.is_some() {
+            return Ok(Event::Sync);
+        }
+
+        if self.input.has_data() {
+            let mut data_block = self.input.pull_data().unwrap()?;
+            if let Some(block_meta) = data_block
+                .take_meta()
+                .and_then(AggregateMeta::downcast_from)
+            {
+                self.input_data = Some(block_meta);
+                return Ok(Event::Sync);
+            }
+        }
+
+        if self.input.is_finished() {
+            if matches!(self.stage, Stage::Input) {
+                // the stage that get meta from input is end now
+                // begin to get meta from channel
+                self.stage = Stage::Channel;
+                return Ok(Event::Sync);
+            }
+
+            if !self.should_finish {
+                return Ok(Event::Async);
+            }
+            self.output.finish();
+            return Ok(Event::Finished);
+        }
+
+        self.input.set_need_data();
+        Ok(Event::NeedData)
+    }
+
+    fn process(&mut self) -> Result<()> {
+        let input_data = self.input_data.take();
+        if let Some(meta) = input_data {
+            self.ensure_spill_depth(0);
+            self.handle_meta(meta, self.max_aggregate_spill_level > 0)?;
+            return Ok(());
+        } else if let Some(mut task) = self.channel_data.take() {
+            self.ensure_spill_depth(task.spilled_depth);
+            self.statistics.reset();
+            let meta = AggregateMeta::Spilled(mem::take(&mut task.spilled_payload));
+            if task.spilled_depth >= self.max_aggregate_spill_level {
+                self.handle_meta(meta, false)?;
+            } else {
+                self.handle_meta(meta, true)?;
+            }
+            self.finish(Some(task.task_id), task.spilled_depth, task.tx)?;
+
+            return Ok(());
+        } else {
+            self.ensure_spill_depth(0);
+            let sender = mem::take(&mut self.tx)
+                .expect("logic error: called finished for input data more than once");
+            self.finish(None, 0, sender)?;
+        }
+
+        Ok(())
+    }
+
+    #[async_backtrace::framed]
+    async fn async_process(&mut self) -> Result<()> {
+        match self.rx.recv().await {
+            Ok(meta) => {
+                self.channel_data = Some(meta);
+            }
+            Err(_) => {
+                self.should_finish = true;
+            }
+        }
+
+        Ok(())
     }
 }
