+++ tests/sqllogictests/suites/mode/standalone/explain/limit.test
@@ -272,13 +272,13 @@ digraph {
     3 [ label = "CompoundBlockOperator(Map)" ]
     4 [ label = "TransformRankLimitSort" ]
     5 [ label = "TransformRankLimitSort" ]
-    6 [ label = "NewTransformPartialAggregate" ]
-    7 [ label = "NewTransformPartialAggregate" ]
+    6 [ label = "TransformPartialAggregate" ]
+    7 [ label = "TransformPartialAggregate" ]
     8 [ label = "LocalScatterTransform(Bucket)" ]
     9 [ label = "LocalScatterTransform(Bucket)" ]
     10 [ label = "ExchangeShuffleTransform" ]
-    11 [ label = "NewTransformFinalAggregate" ]
-    12 [ label = "NewTransformFinalAggregate" ]
+    11 [ label = "TransformFinalAggregate" ]
+    12 [ label = "TransformFinalAggregate" ]
     13 [ label = "Resize" ]
     14 [ label = "Resize" ]
     15 [ label = "LimitTransform" ]
