+++ tests/sqllogictests/suites/tpch/aggregate_shuffle.test
@@ -6,9 +6,6 @@ statement ok
 use tpch_test;
 
 statement ok
-set enable_experiment_aggregate = 1;
-
-statement ok
 set force_aggregate_shuffle_mode = 'row';
 
 # TPC-H TEST
